Fix problems with bulk inserts that are more than a unique index
Let's say you have a table with rows `foo`, `bar`, and `baz`.  There is a
unique index on `foo` and `bar`.  In your database, there is currently a
row:

    foo=1, bar=2, baz=3

You want to bulk insert a row

    foo=1, bar=2, baz=180

Previously, this attempt would have exploded with a `PG::UniqueViolation`,
because the query we *used* to do would have compared foo, bar, *and* baz
when attempting to filter out "duplicate" rows, and the row we're trying to
insert wouldn't have matched the row in the database, and so the insert
would have been attempted (and would have failed, spectacularly).

Now, I've added some intelligence to the `bulk_insert` call, introspecting
the database to determine what unique index(es) exist, and what columns are
in them.  We then do the does-this-row-exist comparison *only* against those
columns, rather than against the entire database, resulting in no
unnecessary assplosions.